Changed handler file name and add method to cors ops
