+++ socket.io/index.js
@@ -1,10 +1,11 @@
 import { Server } from 'socket.io';
-import { handleConnect } from './handlers/socket.io.js';
+import { handleConnect } from './handlers/handleConnect.js';
 
 
 const io = new Server(5400, {
   cors: {
     origin: "http://localhost:3000",
+    methods: ["GET"]
   }
 });
